requirementsAnterior.txt
#Django==1.10.3
django-braces==1.10.0
django-cors-headers==1.3.1
django-oauth-toolkit==0.10.0
djangorestframework==3.5.3
docutils==0.12
oauthlib==1.0.3
Pillow==3.4.2
PyMySQL==0.7.9
six==1.10.0

# django-backend-utils==0.1 pip install django-backend-utils-0.1.zip
# django-oauth2-backend==0.1 pip install django-oauth2-backend-0.1.zip